backend/requirements.txt
django
djangorestframework
gunicorn
django-heroku
django-cors-headers